Use the storyboard for RootViewController and ContainerViewController to more easily absorb the difference between OS and device.
